Finalizado

+++ Aula12/Controllers/AutenticacoesController.cs
@@ -17,6 +17,8 @@ namespace Aula12.Controllers
         // GET: Autenticacoes
         public ActionResult Index()
         {
+            if (!IsAdmin())
+                return RedirectToAction("index", "Inicio", new { msg = "Você não está autorizado" });
             var autenticacoes = db.Autenticacoes.Include(a => a.Usuario);
             return View(autenticacoes.ToList());
         }
@@ -24,6 +26,8 @@ namespace Aula12.Controllers
         // GET: Autenticacoes/Details/5
         public ActionResult Details(int? id)
         {
+            if (!IsAdmin())
+                return RedirectToAction("index", "Inicio", new { msg = "Você não está autorizado" });
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -39,6 +43,8 @@ namespace Aula12.Controllers
         // GET: Autenticacoes/Create
         public ActionResult Create()
         {
+            if (!IsAdmin())
+                return RedirectToAction("index", "Inicio", new { msg = "Você não está autorizado" });
             ViewBag.UsuarioId = new SelectList(db.Usuarios, "UsuarioId", "Nome");
             return View();
         }
@@ -50,6 +56,8 @@ namespace Aula12.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "AutenticacaoId,Data,UsuarioId")] Autenticacao autenticacao)
         {
+            if (!IsAdmin())
+                return RedirectToAction("index", "Inicio", new { msg = "Você não está autorizado" });
             if (ModelState.IsValid)
             {
                 db.Autenticacoes.Add(autenticacao);
@@ -64,6 +72,8 @@ namespace Aula12.Controllers
         // GET: Autenticacoes/Edit/5
         public ActionResult Edit(int? id)
         {
+            if (!IsAdmin())
+                return RedirectToAction("index", "Inicio", new { msg = "Você não está autorizado" });
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -84,6 +94,8 @@ namespace Aula12.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "AutenticacaoId,Data,UsuarioId")] Autenticacao autenticacao)
         {
+            if (!IsAdmin())
+                return RedirectToAction("index", "Inicio", new { msg = "Você não está autorizado" });
             if (ModelState.IsValid)
             {
                 db.Entry(autenticacao).State = EntityState.Modified;
@@ -97,6 +109,8 @@ namespace Aula12.Controllers
         // GET: Autenticacoes/Delete/5
         public ActionResult Delete(int? id)
         {
+            if (!IsAdmin())
+                return RedirectToAction("index", "Inicio", new { msg = "Você não está autorizado" });
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -114,6 +128,8 @@ namespace Aula12.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            if (!IsAdmin())
+                return RedirectToAction("index", "Inicio", new { msg = "Você não está autorizado" });
             Autenticacao autenticacao = db.Autenticacoes.Find(id);
             db.Autenticacoes.Remove(autenticacao);
             db.SaveChanges();
@@ -127,6 +143,19 @@ namespace Aula12.Controllers
                 db.Dispose();
             }
             base.Dispose(disposing);
+        }
+        public bool IsAdmin()
+        {
+            if (Session["Usuario"] != null)
+            {
+                if ("Administrador" == ((Usuario)Session["Usuario"]).StatusUsuario.NomeStatus)//Verifica se o usuario logado é um adm
+                {
+                    ViewBag.IsAdmin = true;
+                    return true;
+                }
+            }
+            ViewBag.IsAdmin = false;
+            return false;
         }
     }
 }
